✨ feature: implement UserSearchModal for participant addition and update user fetching logic
